Move to loop benchmarking folder
